[CO] get_lam_list simplification
