+++ pyleecan/Classes/MachineDFIM.py
@@ -26,11 +26,6 @@ try:
 except ImportError as error:
     get_machine_type = error
 
-try:
-    from ..Methods.Machine.MachineDFIM.get_lam_list import get_lam_list
-except ImportError as error:
-    get_lam_list = error
-
 
 from ._check import InitUnKnowClassError
 from .LamSlotWind import LamSlotWind
@@ -65,17 +60,6 @@ class MachineDFIM(MachineAsync):
         )
     else:
         get_machine_type = get_machine_type
-    # cf Methods.Machine.MachineDFIM.get_lam_list
-    if isinstance(get_lam_list, ImportError):
-        get_lam_list = property(
-            fget=lambda x: raise_(
-                ImportError(
-                    "Can't use MachineDFIM method get_lam_list: " + str(get_lam_list)
-                )
-            )
-        )
-    else:
-        get_lam_list = get_lam_list
     # save and copy methods are available in all object
     save = save
     copy = copy
