+++ pyleecan/Generator/ClassesRef/Machine/Machine.csv
@@ -14,3 +14,4 @@ logger_name,-,Name of the logger to use,0,str,Pyleecan.Machine,,,,,,comp_masses,
 ,,,,,,,,,,,get_polar_eq,,,
 ,,,,,,,,,,,plot,,,
 ,,,,,,,,,,,plot_anim_rotor,,,
+,,,,,,,,,,,get_lam_list,,,
