+++ pyleecan/Generator/ClassesRef/Machine/MachineDFIM.csv
@@ -1,4 +1,3 @@
 Variable name,Unit,Description (EN),Size,Type,Default value,Minimum value,Maximum value,,Package,Inherit,Methods,Constant Name,Constant Value,Class description
 rotor,,Machine's Rotor,,LamSlotWind,,,,,Machine,MachineAsync,check,VERSION,1,Doubly Fed Induction Machine
 stator,,Machine's Stator,,LamSlotWind,,,,,,,get_machine_type,,,
-,,,,,,,,,,,get_lam_list,,,
