+++ pyleecan/Generator/ClassesRef/Machine/MachineIPMSM.csv
@@ -1,4 +1,3 @@
 Variable name,Unit,Description (EN),Size,Type,Default value,Minimum value,Maximum value,,Package,Inherit,Methods,Constant Name,Constant Value,Class description
 rotor,,Machine's Rotor,,LamHole,,,,,Machine,MachineSync,check,VERSION,1,Interior Permanent Magnet Synchronous Machine
 stator,,Machine's Stator,,LamSlotWind,,,,,,,get_machine_type,,,
-,,,,,,,,,,,get_lam_list,,,
