+++ pyleecan/Generator/ClassesRef/Machine/MachineSIPMSM.csv
@@ -1,4 +1,3 @@
 Variable name,Unit,Description (EN),Size,Type,Default value,Minimum value,Maximum value,,Package,Inherit,Methods,Constant Name,Constant Value,Class description
 rotor,,Machine's Rotor,,LamSlotMag,,,,,Machine,MachineSync,check,VERSION,1,Inset and Surface Permanent Magnet Synchronous Machine
-stator,,Machine's Stator,,LamSlotWind,,,,,,,get_lam_list,,,
-,,,,,,,,,,,get_machine_type,,,
+stator,,Machine's Stator,,LamSlotWind,,,,,,,get_machine_type,,,
