+++ pyleecan/Generator/ClassesRef/Machine/MachineSyRM.csv
@@ -1,4 +1,3 @@
 Variable name,Unit,Description (EN),Size,Type,Default value,Minimum value,Maximum value,,Package,Inherit,Methods,Constant Name,Constant Value,Class description
 rotor,,Machine's Rotor,,LamHole,,,,,Machine,MachineSync,check,VERSION,1,Synchronous Reluctance Machine
-stator,,Machine's Stator,,LamSlotWind,,,,,,,get_lam_list,,,
-,,,,,,,,,,,get_machine_type,,,
+stator,,Machine's Stator,,LamSlotWind,,,,,,,get_machine_type,,,
