+++ pyleecan/Generator/ClassesRef/Machine/MachineUD.csv
@@ -1,5 +1,4 @@
 Variable name,Unit,Description (EN),Size,Type,Default value,Minimum value,Maximum value,,Package,Inherit,Methods,Constant Name,Constant Value,Class description
 lam_list,,List of Lamination,,[Lamination],,,,,Machine,Machine,build_geometry,VERSION,1,User defined Machine with multiple Laminations
-is_sync,,True if the machine should be handled as a Synchronous machine,,bool,True,,,,,,get_lam_list,,,
-,,,,,,,,,,,plot,,,
+is_sync,,True if the machine should be handled as a Synchronous machine,,bool,True,,,,,,plot,,,
 ,,,,,,,,,,,is_synchronous,,,
